moved confirmation modal
