+++ app/react/V2/Components/Forms/index.ts
@@ -2,7 +2,6 @@ export { InputField } from './InputField';
 export { Select } from './Select';
 export { MultiSelect } from './MultiSelect';
 export { RadioSelect } from './RadioSelect';
-export { ConfirmNavigationModal } from './ConfirmNavigationModal';
 export { Label } from './Label';
 export { Geolocation } from './Geolocation';
 export { FileDropzone } from './FileDropzone';
