+++ app/react/V2/Components/UI/index.ts
@@ -16,6 +16,7 @@ export { PaginationState } from './PaginationState';
 export { MediaPlayer } from './MediaPlayer';
 export { FileIcon } from './FileIcon';
 export { NeedAuthorization } from './NeedAuthorization';
+export { ConfirmNavigationModal } from './ConfirmNavigationModal';
 
 export type { PillColor } from './Pill';
 export type { TableProps, TableRow } from './Table';
