+++ app/react/V2/Routes/Settings/CustomUploads/CustomUploads.tsx
@@ -9,9 +9,8 @@ import { Translate } from 'app/I18N';
 import { FetchResponseError } from 'shared/JSONRequest';
 import { FileType } from 'shared/types/fileType';
 import { getByType, remove, UploadService } from 'V2/api/files';
-import { Button, ConfirmationModal, Table } from 'V2/Components/UI';
+import { Button, ConfirmationModal, Table, ConfirmNavigationModal } from 'V2/Components/UI';
 import { SettingsContent } from 'V2/Components/Layouts/SettingsContent';
-import { ConfirmNavigationModal } from 'V2/Components/Forms';
 import { notificationAtom } from 'V2/atoms';
 import {
   createColumns,
