+++ app/react/V2/Routes/Settings/MenuConfig/MenuConfig.tsx
@@ -8,10 +8,9 @@ import { useSetAtom } from 'jotai';
 import { cloneDeep, isEqual } from 'lodash';
 import { Translate } from 'app/I18N';
 import * as SettingsAPI from 'app/V2/api/settings';
-import { ConfirmNavigationModal } from 'app/V2/Components/Forms';
 import { notificationAtom } from 'app/V2/atoms';
 import { settingsAtom } from 'app/V2/atoms/settingsAtom';
-import { Button, Table, Sidepanel } from 'app/V2/Components/UI';
+import { Button, Table, Sidepanel, ConfirmNavigationModal } from 'app/V2/Components/UI';
 import { SettingsContent } from 'app/V2/Components/Layouts/SettingsContent';
 import { MenuForm } from './components/MenuForm';
 import { columns } from './components/TableComponents';
