+++ app/react/V2/Routes/Settings/Pages/PageEditor.tsx
@@ -19,9 +19,9 @@ import { Translate, t } from 'app/I18N';
 import * as pagesAPI from 'V2/api/pages';
 import { Page } from 'V2/shared/types';
 import { SettingsContent } from 'V2/Components/Layouts/SettingsContent';
-import { Button, CopyValueInput, Tabs } from 'V2/Components/UI';
+import { Button, CopyValueInput, Tabs, ConfirmNavigationModal } from 'V2/Components/UI';
 import { CodeEditor } from 'V2/Components/CodeEditor';
-import { ConfirmNavigationModal, EnableButtonCheckbox, InputField } from 'app/V2/Components/Forms';
+import { EnableButtonCheckbox, InputField } from 'app/V2/Components/Forms';
 import { notificationAtom } from 'V2/atoms';
 import { FetchResponseError } from 'shared/JSONRequest';
 import { getPageUrl } from './components/PageListTable';
