+++ app/react/V2/Routes/Settings/Thesauri/EditThesaurus.tsx
@@ -8,8 +8,7 @@ import { isEmpty } from 'lodash';
 import { Translate } from 'app/I18N';
 import { ClientThesaurus } from 'app/apiResponseTypes';
 import { SettingsContent } from 'V2/Components/Layouts/SettingsContent';
-import { Button } from 'V2/Components/UI';
-import { ConfirmNavigationModal } from 'V2/Components/Forms';
+import { Button, ConfirmNavigationModal } from 'V2/Components/UI';
 import { notificationAtom, templatesAtom } from 'app/V2/atoms';
 import { ThesaurusSchema } from 'shared/types/thesaurusType';
 import { PropertySchema } from 'shared/types/commonTypes';
