+++ app/react/V2/Routes/Settings/Translations/EditTranslations.tsx
@@ -17,10 +17,10 @@ import { useSetAtom } from 'jotai';
 import { Translate } from 'app/I18N';
 import { advancedSort } from 'app/utils/advancedSort';
 import { ClientTranslationSchema } from 'app/istore';
-import { ConfirmNavigationModal, InputField } from 'app/V2/Components/Forms';
+import { InputField } from 'app/V2/Components/Forms';
 import { SettingsContent } from 'app/V2/Components/Layouts/SettingsContent';
 import RenderIfVisible from 'react-render-if-visible';
-import { Button, ToggleButton } from 'V2/Components/UI';
+import { Button, ToggleButton, ConfirmNavigationModal } from 'V2/Components/UI';
 import * as translationsAPI from 'V2/api/translations';
 import * as settingsAPI from 'V2/api/settings';
 import { notificationAtom } from 'V2/atoms';
